Provide a way to install Python on Bash and Zsh on Linux and MacOS
The script to realize this expects Git to be installed so that it
can downlod pyenv from GitHub, and install the latest available
Python or one close to it through pyenv.

The user of the script will be asked to add the locations of Python
to PATH so that installation can be completed and Python subsequently
found for other purposes such as for installing the EBCLI.

As an aside, the script also installs virtualenv so that the EBCLI
self-contained installer can use it. For this, the "pip" executable
located within the just-installed Python's "bin" directory
(/path-to/pyenv/versions/<PYTHON_VERSION>/bin/) is used.