Avoid closing existing shell after installation completion
Originally, opening up a new PowerShell session solved the problem of
changes to $env:PATH not taking effect in the current session. However,
this can be solved by updating the current session's PATH variable.

`Cleanup` is no longer necessary, so the only remaining statement (to
remove the downloaded Python MSI) is extracted out. Additionally, this
step is performed only if the Python MSI was downloaded.