Install Python 3.7.2 on Unix/Linux
As of release 1.2.9, pyenv does not support installation of Python
beyond 3.7.2 on Linux. This is not a problem on OS X as 3.7.3 can be
installed. However, since both OS X and Linux run the same
bundled_installer script, we need to lower the version to install on
both to 3.7.2 for the time being until we can add logic to the script
to distinguish between Darwin and Linux.