Verify a valid HTTP client tool exists prior to installing Python
On Linux, Python installation requires that the Python code be downloaded.
In order to download Python, pyenv uses one of aria2c, curl, and wget.
When none of these is found, it fails to download Python.

This commit ensures that python_installer fast-fails so that
unnecessary activities such as cloning the pyenv repository need not
happen when a valid HTTP client tool is not found.