Fix issue involving spaces within virtualenv DEST_DIR path
When the value of the --location argument of ebcli_installer.py (or
the path of the HOME directory) contain spaces, the installer passes
multiple values in place of virtualenv's DEST_DIR positional argument.
virtualenv rejects the multi-value input and causes the creation of
the EB CLI's vrtualenv to fail.

This was originally reported in Issue #20 [1].

This commit fixes the problem by enclosing the value for DEST_DIR within
quotes before passing it to the virtualenv executable.

[1] https://github.com/aws/aws-elastic-beanstalk-cli-setup/issues/20