Default to creating virtualenv using Python used to invoke ebcli_installer.py
The current behaviour of the ebcli_installer.py is to not pass a Python
executable's PATH to virtualenv to create a sandbox for the EB CLI.
As a result, virtualenv may use a Python other than the one used to
invoke ebcli_installer.py. This behaviour is confusing. On some systems
where the default Python is 2.7, virtualenv may use that, thereby
defeating the goal to use the latest and greatest.

This commit ensures that ebcli_installer.py passes the path to the
Python executable used to invoke it to virtualenv to create a sandbox
for the EB CLI with.